backend/requirements.txt
# --- FastAPI & Core ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.27.0
requests==2.32.3
pydantic==2.5.3
pydantic-settings==2.2.1

# --- Database ---
sqlalchemy==2.0.25

# --- AI/ML Stack ---
numpy==1.26.4
tensorflow==2.16.1
tf-keras==2.16.0
deepface==0.0.92
opencv-python==4.8.1.78
pillow==10.1.0

# --- Security & Auth ---
openai==1.52.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- New Additions ---
motor==3.3.2
pymongo==4.6.1
bcrypt
sentence-transformers
chromadb
email-validator
geopy
groq